coalesce the store_ingest and builder servers
The nix_store server now performs the functions of both the
store_ingest and builder servers. This moves all file system
trickry to one component.

The build server required all file system sessions other than the
output session to be rooted to an input path. The nix_store now has
a filter service for build children that allows all inputs to be
served from a single top-level store session while denying access
to files and directories not declared in the derivation.